e2e/tests/requirements.txt
orbs-orderbook-sdk==0.9.2
pytest==7.4.4
redis==5.0.3